ref_gl: fix crash bug on lightmaps with high-magnitude scales.
These lightmaps are generated by a special qrad3 feature that can shrink
the lightmap textures for surfaces that are completely uniform in lighting
color. Turns out qrad3 was generating values that caused integer overflows in
rare circumstances. Solution is to increase the integer size used for these
calculations.

+++ trunk/source/ref_gl/r_model.c
@@ -943,7 +943,7 @@ void CalcSurfaceExtents (msurface_t *s, qboolean override, int *smax, int *tmax,
 	int		i,j, e, vnum;
 	mvertex_t	*v;
 	mtexinfo_t	*tex;
-	int		bmins[2], bmaxs[2];
+	int64_t bmins[2], bmaxs[2];
 
 	mins[0] = mins[1] = 999999;
 	maxs[0] = maxs[1] = -99999;
